requirements.txt
Jinja2==2.6
Pygments==1.6
Sphinx==1.2b1
docutils==0.10
wsgiref==0.1.2
PyOpenGL==3.0.2
PyOpenGL-accelerate==3.0.2
